pyproject.toml
[project]
name = "balance-bot-2"
version = "0.1.0"
description = "Self-balancing robot homebrew toy"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mpu6050-raspberrypi>=1.2",
    "pydantic>=2.12.5",
    "simple-pid>=2.0.1",
    "PyGLM>=2.7.0",
    "smbus2",
]

[project.scripts]
balance-bot = "balance_bot.main:main"

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "pytest>=9.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["balance_bot"]
